# runtime/internal/openapidocs/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "mcpserverapi_oauthStartRequest": {
        "properties": {
          "redirectBase": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "mcpserverapi_oauthStartResponse": {
        "properties": {
          "authorizationUrl": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "runtimetypes_MCPServer": {
        "properties": {
          "args": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "authEnvKey": {
            "type": "string"
          },
          "authToken": {
            "type": "string"
          },
          "authType": {
            "type": "string"
          },
          "command": {
            "type": "string"
          },
          "connectTimeoutSeconds": {
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "id": {
            "type": "string"
          },
          "injectParams": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "name": {
            "type": "string"
          },
          "oauthClientId": {
            "type": "string"
          },
          "oauthClientSecretEnv": {
            "type": "string"
          },
          "transport": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API served by `contenox serve` under /api. Generated from route annotations by tools/openapi-gen.",
    "title": "Contenox Runtime API",
    "version": "v0.36.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/mcp-servers": {
      "get": {
        "operationId": "mcpserver_list",
        "parameters": [
          {
            "description": "An optional RFC3339Nano timestamp to fetch the next page of results.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The maximum number of items to return per page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/runtimetypes_MCPServer"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "list returns the registered MCP servers, paginated by cursor.",
        "tags": [
          "mcpserver"
        ]
      },
      "post": {
        "operationId": "mcpserver_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/runtimetypes_MCPServer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/runtimetypes_MCPServer"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "create registers an MCP server and returns it with its assigned ID.",
        "tags": [
          "mcpserver"
        ]
      }
    },
    "/mcp-servers/by-name/{name}": {
      "get": {
        "operationId": "mcpserver_getByName",
        "parameters": [
          {
            "description": "The unique name of the MCP server.",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/runtimetypes_MCPServer"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "getByName returns one MCP server by its unique name.",
        "tags": [
          "mcpserver"
        ]
      }
    },
    "/mcp-servers/{id}": {
      "delete": {
        "operationId": "mcpserver_delete",
        "parameters": [
          {
            "description": "The unique ID of the MCP server.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "delete removes an MCP server registration by ID.",
        "tags": [
          "mcpserver"
        ]
      },
      "get": {
        "operationId": "mcpserver_get",
        "parameters": [
          {
            "description": "The unique ID of the MCP server.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/runtimetypes_MCPServer"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "get returns one MCP server by ID.",
        "tags": [
          "mcpserver"
        ]
      },
      "put": {
        "operationId": "mcpserver_update",
        "parameters": [
          {
            "description": "The unique ID of the MCP server.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/runtimetypes_MCPServer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/runtimetypes_MCPServer"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "update replaces an MCP server's configuration by ID and returns the stored result.",
        "tags": [
          "mcpserver"
        ]
      }
    },
    "/mcp-servers/{id}/oauth/start": {
      "post": {
        "operationId": "mcpserver_oauthStart",
        "parameters": [
          {
            "description": "The unique ID of the MCP server.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/mcpserverapi_oauthStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/mcpserverapi_oauthStartResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "oauthStart begins the OAuth authorization flow for an MCP server and returns the provider authorization URL to visit.",
        "tags": [
          "mcpserver"
        ]
      }
    },
    "/mcp/oauth/callback": {
      "get": {
        "operationId": "mcpserver_oauthCallback",
        "parameters": [
          {
            "description": "The OAuth authorization code returned by the provider.",
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The provider error code, when authorization fails.",
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The provider error description, when authorization fails.",
            "in": "query",
            "name": "error_description",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The OAuth state value returned by the provider.",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "302 redirect back to the client UI with success or error query parameters appended; no response body."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "oauthCallback completes an MCP server OAuth flow from the provider's redirect and 302-redirects back to the client UI.",
        "tags": [
          "mcpserver"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}